Fix race conditions on data transfer by implementing the possibility of enqueuing a request to DataStream that is only started after an event. This commit compliments 9531631fd3a58a0d40fd37d8d813d71dc398c5fd and the main case that it solves is the one exposed by standalone_tests/test_createbuffer_ptr_race.
--

In this test, many enqueueWriteBuffer calls which also require a synchronization operation
(downloading an OpenCL buffer) are enqueued.

This caused a race because the synchronization operation sended the message that started the
transfer on the host (EventSynchronizationMessage) immediately while the command was enqueued,
but didn't start the data transfer on the ode until when some OpenCL commands were completed.

On the other hands, the write buffer did send both the command to start the transfer on
the host (CommandExecutionStatusChangedMessage(CL_SUBMITTED)) and the data transfer on the node
only after a callback when some OpenCL commands were completed.

This caused the following race to become possible, when two enqueueWriteBuffer with
both requiring a buffer acquire synchronization operation were emitted in succession:
* First enqueueWriteBuffer sends EventSynchronizationMessage to the host
* First enqueueWriteBuffer enqueues synchronization operations on the node
* First enqueueWriteBuffer enqueues write buffer command on the node
* Second enqueueWriteBuffer sends EventSynchronizationMessage to the host
* Second enqueueWriteBuffer enqueues synchronization operations on the node
* Second enqueueWriteBuffer enqueues write buffer command on the node
* First enqueueWriteBuffer synchronization operations are executed
      * Host side's EventSynchronizationMessage matches synchronization on node --> OK
* First enqueueWriteBuffer write buffer command is executed
      * Node sends a CommandExecutionStatusChangedMessage(CL_SUBMITTED) to the host
      * But, a EventSynchronizationMessage was already sent to the host before, so the
        host's side EventSynchronizationMessage matches enqueueWriteBuffer on node --> KO

--

The way this is solved in this commit is by pairing all messages that cause a data transfer in
the host (EventSynchronizationMessage, CommandExecutionStatusChangedMessage(CL_SUBMITTED), etc.)
to immediately enqueuing the data transfer on the DataStream, to guarantee that all data transfer
pairs match correctly on both the host side and the node side.

Since often a data transfer can not be started immediately on the host or node, the way this works
is that even though the transfer is enqueued (so it will delay any successive transfers), the start
of the operation can be delayed until an event completes, so the transfer can only start when its
dependencies are satisfied.

At first sight, this may seem to introduce unnecessary serialization which could slow down execution,
but I believe that with the current design, all "un-serialized" operations were actually matching
the wrong transfer sides on both the node and the host, so they were essentially race conditions.
(It would certainly be interesting to eventually have a more asynchronous design where newer data
transfers operations could be executed before old ones if they are independent, but I believe that
that would be useless anyway unless out-of-order OpenCL queues were implemented)

--

This commit also fixes another "simpler" race condition where in CLComputeNodeEventProcessor,
executionStatusChanged pushed to the _taskList but RequestBufferTransfer didn't,
which could cause those commands to be executed out-of-order in the host even though they arrived
in the correct order.

This commit is missing some error checking logic that was present before for exceptional cases
(to be added later).