CLRequestProcessor.cpp: Fix build in C++14/C++17 mode.
While not an immediate problem, since dOpenCL is still built in C++11 mode, trying to compile it in C++14 (or greater) mode caused a compilation error due to the make_unique function in CLRequestProcessor (used as a workaround for the lack of std::make_unique in C++11).

Despite this workaround (or 'polyfill') being defined in the global namespace (::make_unique) and not on the std namespace (std::make_unique), it still caused an ambiguous overload error on new C++ versions, due to argument-dependent name lookup ( https://en.wikipedia.org/wiki/Argument-dependent_name_lookup ).

To fix this, place our implementation its own named namespace to ensure we always unambiguously call our implementation for now.