dclasio/src/dcl/CLObjectRegistry.cpp: Make CLObjectRegistry thread-safe to avoid undefined behaviour / hangs.
When running the MemoryConsistency/ConcurrentWriteRead test a few dozen / hundred times, I could consistently reproduce a hang. I finally traced this to CLObjectRegistry not being thread safe.

CLObjectRegistry can be accessed concurrently from both the main thread and the worker thread (in the case of the above test, the detail::Registry<CommandListener> is accessed both from the main thread on  _cl_command_queue::enqueueRead on icdpp/src/CommandQueue.cpp, and from the worker thread on CLComputeNodeEventProcessor::executionStatusChanged on dclasio/src/dclasio/comm/CLEventProcessor.cpp). Since C++ STL collections are not thread safe, this causes undefined behavior.

On my case, this undefined behavior manifested as the lookup (in CLComputeNodeEventProcessor::executionStatusChanged) failing when it shouldn't have, causing a tricky deadlock later on.

The fix is simply to introduce a mutex.