(daemon|icdpp)/src/CommandQueue.cpp: Fix race between certain commands and synchronization events.
This fix is similar to other previous fixes for data races between messages between the host and the nodes. Basically, the send/receives between the host and the node for commands related to data/memory transfers (enqueueReadBuffer/enqueueWriteBuffer/enqueueMapBuffer/enqueueUnmapBuffer) could get mismatched due to data races.

In this particular case, this data race could be reproduced pretty consistently by running the MemoryConsistency/ConcurrentWriteRead test a few dozen times. Eventually, the data transfer send from node -> host for returning the data from the enqueueReadBuffer and the data transfer from node -> host related to the buffer synchronization with the *other* enqueueReadBuffer requested by the other node get exchanged, causing the results (vecOut0 and vecOut1 in the test) to get exchanged wrt. the expected result.

Like other similar fixes in the past, this happens because in the communication between the host->node, the message and the data transfer for the command were enqueued at different times (in this case, after receiving the message for the command, the node sent a message back to the host to set the status to CL_SUBMITTED; starting the transfer). In this 'gap' between the message and the data transfer, a synchronization event (from another node) can 'sneak in between' and get mismatched.

The fix in this commit is simply to enqueue both the command and the data transfer at the same time to close this gap, so they always get matched correctly.

Like other fixes, this makes communication more 'synchronous', but avoids the problem. The more I think about it, the more I believe the original intention of the authors of dOpenCL was for transfers to be completely asynchronous, so they could be executed in a different order than enqueued. However, this was never actually implemented (e.g. DataStream had no way to run transfers in a different order than enqueued, and the transfers don't even have something like a unique ID so they could be 'matched' with the right transfers on to the other side). For now my priority with this fixes is getting correctness and 'simple things working', but actually implementing this would be good for performace for applications with a lot of communication, and also likely necessary for eventually getting out of order queues working.