Make dcl::util::Logger thread safe to avoid undefined behavior
Before this commit, dcl::util::Logger was not thread safe (despite using a mutex to synchronize writing to the output stream, the logger itself was not thread safe, because it inherited std::ostream which is not thread safe). This commit makes it thread safe.

For this, the logger has been refactored to have a thread-safe core (LoggerImpl), along with a non intrinsically thread safe formatting interface (LoggerStreamImpl), which is then used with thread local variables in order to make it thread safe.

A test is also added.

This commit also includes some other small refactors around the logger interface, such as e.g. removing unused methods (LoggerImpl::setDefaultPriority), fixing Clang-Tidy warnings, making more methods private, etc.

After this commit, we should both avoid the undefined behavior of the logger implementation on multithreading, as well as the crashes on process termination with Boost.Log.