Document Boost.Thread dependency and install in CI and Docker environments.
This dependency is required by the use of some boost::thread features in dclasio (e.g. boost::barrier in DataStream). It was not properly documented, but build worked anyway because it is was a dependency of Boost.Log. However, since 96667ac1e31522871123711f578dc5a6488db57d 'Recover the old logger, which was replaced by Boost.Log.', this doesn't happen anymore. Document and install it properly from now on.