dclasio/src/dclasio/comm/DataStream.cpp: Fix race condition on send/compress path for I/O link compression.
When I/O link compression is enabled, the send/compress path is called from the compression threads (i.e. from DataStream::loop_compress_thread). This can cause 'spurious' calls that should be ignored. The case where a call should be ignored because a write was already in progress was already handled, but the case where the last send had already been completed was not.

While this race condition was very rare, when it happened, it caused undefined behavior manifested in hangs, use-after-free, etc..

The fix is to manipulate _write_io_num_blocks_remaining always inside a lock and set a special value that tells it to ignore a call once a send operation has finished. The _write_io_total_bytes_transferred variable is also now inside the lock because it could otherwise be updated by multiple threads in very unlucky race cases.