Update OpenCL C++ bindings: cl.hpp -> cl2.hpp.
The cl.hpp header from the C++ binding is deprecated and it's quite trivial to update to cl2.hpp. cl2.hpp is also available at least on Ubuntu 16.04+ and Debian stretch+ so it should be everywhere now (or can be easily downloaded otherwise).

Most changes for the upgrade are quite trivial:
* Change include of CL/cl.hpp -> CL/cl2.hpp
* Rename __CL_ENABLE_EXCEPTIONS -> CL_HPP_ENABLE_EXCEPTIONS
* Rename VECTOR_CLASS -> cl::vector
* #define CL_HPP_MINIMUM_OPENCL_VERSION 120 and CL_HPP_TARGET_OPENCL_VERSION 120 (otherwise defaults to 2.0 and complains)

Some more tricky changes are:
* (On daemon/src/Context.cpp) nullptr is not implicitly casted to cl::Buffer anymore (must construct explicitly
* (On daemon/src/Kernel.cpp) Kernel parameter must be set as cl::Memory and not cl::Memory*
* (On daemon/src/Program.cpp) Program sources must be std::string and not (char*,size_t) pair.

Additionally, for the CI builds, on Ubuntu 16.04 cl.hpp is not available, but cl2.hpp is, so that's one problem less to care about.