Implement out-of-order transfer matching for DataStream.
This commits allows transfers in DataStream to be executed in a different order than submitted. Previously, the order of DataStream::read and DataStream::write calls determined the order in which the transfers were executed.

On one side, this can easily lead to deadlocks or mixups if one is not careful about making sure to request the transfers to DataStream in the exact same order on the sender and receiver sides. One example of this is the last commit, 60659889fe5a46829a6ae47d76c92d556d26575b, which needed to change the place where the transfer calls for some commands (enqueue(Map|Unmap|Read|Write)Buffer) happens in order to ensure this.

However, it turns out that that fix was still flawed. It fixed one of the problems (a deadlock in the Memory/MemoryConsistency/ConcurrentReadWrite test) test but introduced another one (a mixup in the test_createbuffer_ptr_race standalone_test, detectable when running in debug and with the USE_SENTINELS macro enabled).

After examining the cause, it's clear that the transfers in DataStream _must_ be able to be execute out-of-order in order to avoid deadlocks. There's a relatively simple case that proves this: Imagine a setup with 3 nodes (N1, N2, N3). Now, both N2 wants to execute a kernel that accesses a buffer that is stored on N1 and synchronize by the OpenCL guarantees (through an event). Concurrently, N3 also wants to execute a kernel that accesses a *different* buffer that is also stored on N1 and synchronize by the OpenCL guarantees. In the current dOpenCL model, both N2 and N3 will ask the host to ask N1 to transfer the buffers. So, after this, the host will send two synchronization requests ('send me this buffer') to N1 and start two receives. But those two synchronization requests from the host to N1 can arrive in any order, since they go through the network. So, N1 can start any of the two sends first, and before this commit, the host has no way to match the correct one.

This commit adds a 'transfer ID' (just a UUID) associated with each operation that requires a transfer (e.g. create program, read buffer, buffer synchronization, etc.). This ID is generated (random UUID) by the 'originator' side of the operation and transferred (along with the operation/command, through the message queue) to the other side. Each side then asks DataStream to perform its side of the transfer (send/receive) by specifying the 'transfer ID' along with the rest of the parameters of the transfer. Then, DataStream internally 'enqueues' the requests and executes them out-of-order, as they are available.

Despite being out-of-order, the order of execution is still guaranteed correct by the rest of the dOpenCL infrastructure (importantly, note that DataStream::read() and DataStream::write() can be provided a 'completable' to control readiness).
The actual order (implementation detail) is that the transfers are done in the order the writes become available (and: a write that becomes available earlier can block a later write, if the reader side is not present). I belive this should be enough to handle all cases with some care, and of course is sufficient to correctly execute all tests we have for now.

Performance-wise, the infrastructure for out-of-order matching is bound to have some overhead, but should be easily offset by actually being able to avoid some blocking by running out-of-order in most cases.

Reverts: 60659889fe5a46829a6ae47d76c92d556d26575b