Remove hacky and test code.
This removes some code that was introduced in the first revisions after deriving this fork from the original SVN r1819 which looks like test code or hacky code used during testing. Possibly some this code was also used for initial integration with CloudCL, so it will be reintroduced if required if problems arise. The tests still run without this code.