dclasio/src/dclasio/comm/DataStream.cpp: Split reads/writes in blocks of size SUPERBLOCK_MAX_SIZE (required for OpenCL decompression) at enqueue time rather than at DataStream::(read|write) call time.
This is still very hacky, but it fixes a race condition in test_createbuffer_ptr_multi when SUPERBLOCK_MAX_SIZE=2**23 (or similar). This race condition happened because by splitting the reads/writes at DataStream::(read|write) call time, it was not guaranteed that each part of a split transfer would be executed one after the other (i.e. other transfers could sneak in between). This violated some assumptions on the OpenCL decompression code and lead to a hang due to a race condition. By splitting the transfers at enqueue time, we can guarantee that they will be executed in the desired order. In reality the only relevant change is the one for the write path, but the read path has also been changed for consistency.