Implement cl(Retain|Release)Device for root-level devices
clRetainDevice and clReleaseDevice is just a no-op for root-level devices,
which are the only device types supported for now by dOpenCL.

This fixes the execution of the standalone_test's when dOpenCL is built in
Debug mode (so assertions trigger). This was first broken when updating to
the new cl2.hpp header, which calls cl(Retain|Release)Device when getting
the list of root-level devices.