Fix a race condition in dcl::BlockingQueue / dclasio::comm::CLComputeNodeEventProcessor
This fixes a race condition that caused a hang that could happen on the creation of very short-lived dclasio::comm::CLComputeNodeEventProcessor instances.
On this case, the destructor of dclasio::comm::CLComputeNodeEventProcessor could be called before the worker thread had actually spawned and was waiting for tasks in the task queue. The destructor interrupted the task queue in order to ask the worker thread to exit and then waited for the worker thread to actually exit. However, when the worker thread actually spawned, the interrupt flag the destructor previously set on the task queue was cleared so the interrupt never had any effect. This led to the destructor of dclasio::comm::CLComputeNodeEventProcessor waiting forever.
The fix in this case is to keep the interrupt flag even if no one is waiting at that exact moment, so the interrupt is eventually delivered when someone does and the thread always exit.

This was reproducible with very short lived unit tests (e.g. test/Platform/GetPlatformInfo) while at the same time high CPU load (to make the thread spawn slow).