Synchronize with the start of remote events with no buffer dependencies + fix bitonic standalone_test
Surprisingly, until now when enqueuing a command with a wait list containing an event from another remote device that didn't have any buffer dependencies, dOpenCL didn't wait for that event to complete before starting the command, as required. I am not sure if this never happened or it got lost in previous changes somehow. Anyway, this fixes the problem, which along with a few other fixes, makes the bitonic sort standalone_test work again.