Remove remaining unnecessary references to Boost.Thread
Boost.Thread was no longer necessary (it was removed in commit 4968e9a74003cace3a09a65c367589d1b65b3634 'Remove dependency on boost::barrier (and hence on the entire boost::thread) by providing a standalone implementation of it.'), but it was not removed from the build scripts, Dockerfiles, and CI scripts. This completes the removal of Boost.Thread.