dclasio/src/dclasio/comm/InputDataStream.cpp: Fix handling of incomplete compresed blocks with OpenCL based decompression.
This caused problems and was detected thanks to standalone_test/test_createbuffer_ptr_multi, which creates a buffer (buf4) with an incomplete block (with the default parameters). Surprisingly this didn't cause problems on Intel iGPUs, but it does on pocl, which is how it got detected.

This should fix the CI build.